crypto: move crypto::Algorithm and methods to crypto/mod.rs
